Move a RSS structure
